refactor: extract Android upload planning
